Move newStripeCheckout into newPaymentFlow subdir
