+++ assets/pages/new-contributions-landing/components/ContributionForm.jsx
@@ -14,7 +14,7 @@ import { type PaymentHandler, type PaymentMethod } from 'helpers/checkouts';
 import { config, type Contrib, type Amount } from 'helpers/contributions';
 import { type CheckoutFailureReason } from 'helpers/checkoutErrors';
 import { emailRegexPattern } from 'helpers/checkoutForm/checkoutForm';
-import { openDialogBox } from 'helpers/paymentIntegrations/newStripeCheckout';
+import { openDialogBox } from 'helpers/paymentIntegrations/newPaymentFlow/newStripeCheckout';
 import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRevenueApis';
 import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import { getAbsoluteURL } from 'helpers/url';
