+++ assets/pages/new-contributions-landing/components/ContributionPayment.jsx
@@ -11,7 +11,7 @@ import { type Contrib } from 'helpers/contributions';
 import { classNameWithModifiers } from 'helpers/utilities';
 import { type IsoCountry } from 'helpers/internationalisation/country';
 import { type IsoCurrency } from 'helpers/internationalisation/currency';
-import { setupStripeCheckout } from 'helpers/paymentIntegrations/newStripeCheckout';
+import { setupStripeCheckout } from 'helpers/paymentIntegrations/newPaymentFlow/newStripeCheckout';
 import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRevenueApis';
 import SvgNewCreditCard from 'components/svgs/newCreditCard';
 import SvgPayPal from 'components/svgs/paypal';
